refactor: 객체의 책임에 따라 domain 과 utils 분리
